Improve Index Creation and Refactoring
If the properties of a newly created index are not
created in the same transaction, sometimes the newly
created index would get into `INSTALLED` state and
never get into `REGISTER_INDEX`. In order to avoid this
issue, all of the running transactions would be rollback
except the one who is used to create the index.

Also add an utility function update composite index's state:
    updateCompositeIndexState()

This could be used to change transite the `REGISTER_INDEX` to
`ENABLED` state.

BTW, Some classes name are changed.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>